python/requirements.txt
asgiref==3.8.1
Django==5.2
sqlparse==0.5.3
psycopg2
djangorestframework
django-cors-headers